moving onSubmitHandler back to CreateArticle component
