+++ src/components/CreateArticle.jsx
@@ -1,12 +1,25 @@
-import React from 'react'
+import React, { useState }  from 'react'
+import axios from "axios";
 import WritingForm from './WritingForm'
 import { Container } from 'semantic-ui-react'
 
 const CreateArticle = () => {
+  const [message, setMessage] = useState("");
+
+  const onSubmitHandler = async (e) => {
+    try {
+      const response = await axios.post("/articles", {
+        params: { title: e.target.title.value, body: e.target.body.value },
+      });
+      setMessage(response.data.message);
+    } catch (error) {
+      setMessage(error.response.data.message);
+    }
+  };
 
   return (
     <Container className="writing-container">
-      <WritingForm />
+      <WritingForm onSubmitHandler={onSubmitHandler} message={message}/>
     </Container>
   );
 };
